feat: add escape button listener in popup
